WIP: Avances en DTOs (Stock, Warehouse, Family)

+++ src/main/java/com/appTest/store/dto/sale/SaleSummaryByDateDTO.java
@@ -4,17 +4,18 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
 @Getter @Setter
 public class SaleSummaryByDateDTO implements Serializable {
     private LocalDate date;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private Long totalSales; // <- Acá el fix
 
     public SaleSummaryByDateDTO() {}
-    public SaleSummaryByDateDTO(LocalDate date, Double totalAmount, Long totalSales) {
+    public SaleSummaryByDateDTO(LocalDate date, BigDecimal totalAmount, Long totalSales) {
         this.date = date;
         this.totalAmount = totalAmount;
         this.totalSales = totalSales;
